pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opengpu-tools"
version = "0.1.0"
description = "Tooling (assembler codegen, ISA spec, runtime helpers) for the OpenSource-GPU-Architecture project."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Piyush Tiwari" },
    { name = "Adam Majmudar" },
]
keywords = ["gpu", "gpgpu", "systemverilog", "simt", "cocotb", "openlane", "yosys"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
]
dependencies = [
    "cocotb>=2.0,<3.0",
    "cocotb-bus>=0.2.1",
    "cocotbext-axi>=0.1.24",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "pytest-cov>=5.0",
    "ruff>=0.6",
    "mypy>=1.10",
    "types-PyYAML",
    "pip-audit>=2.7",
    "pre-commit>=3.7",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocstrings[python]>=0.25",
    "mike>=2.1",
]
ml = [
    "numpy>=1.26",
    "tinygrad>=0.9",
    "safetensors>=0.4",
]

[project.urls]
Homepage = "https://github.com/piyushptiwari1/OpenSource-GPU-Architecture"
Repository = "https://github.com/piyushptiwari1/OpenSource-GPU-Architecture"
Issues = "https://github.com/piyushptiwari1/OpenSource-GPU-Architecture/issues"
Documentation = "https://piyushptiwari1.github.io/OpenSource-GPU-Architecture/"

[project.scripts]
opengpu-asm = "opengpu_tools.asm:main"
opengpu-isa = "opengpu_tools.isa_codegen:main"

[tool.setuptools.packages.find]
where = ["tools"]
include = ["opengpu_tools*"]

# ---------------------------------------------------------------------------
# Ruff (formatter + linter; replaces black + flake8 + isort)
# ---------------------------------------------------------------------------
[tool.ruff]
line-length = 100
target-version = "py310"
# Files under `test/` were inherited from upstream + 15 merged PRs and
# have not yet been reformatted to the new house style. They are
# excluded from ruff format/check enforcement until a dedicated
# "reformat legacy tests" PR lands.
extend-exclude = [
    "build",
    "sim_build",
    "site",
    "third_party",
    "test",
    "main.py",
    "server.py",
]

[tool.ruff.lint]
select = [
    "E", "W",   # pycodestyle
    "F",        # pyflakes
    "I",        # isort
    "B",        # bugbear
    "UP",       # pyupgrade
    "SIM",      # simplify
    "RUF",      # ruff-specific
    "PL",       # pylint subset
    "PTH",      # pathlib
    "S",        # bandit (security)
]
ignore = [
    "S101",     # asserts ok in tests
    "PLR0913",  # too many args (cocotb tests legitimately have many)
]

[tool.ruff.lint.per-file-ignores]
"test/**" = ["S101", "PLR2004"]
"tools/codegen/**" = ["S701"]   # jinja2 autoescape not always needed for code gen
# DiffTest invokes the trusted refsim binary; cmd is constructed from
# typed args, not user input.  Magic small ints in tests are intentional
# (number of records compared).
"tools/difftest/__init__.py" = ["S603"]
"tools/difftest/tests/**" = ["PLR2004"]
# bench/run_all.py drives the trusted in-tree refsim with paths
# constructed from the workspace; not user input. Same exemption applies
# to the tinygrad-opengpu end-to-end test which invokes the same binary.
"bench/run_all.py" = ["S603", "PLW2901"]
"tools/tinygrad_opengpu/tests/**" = ["S603", "PLR2004"]
"tools/asm/tests/**" = ["PLR2004"]
# The assembler is a single-pass translator: a switch on instruction type
# makes more sense than a strategy class for 4 op types. Magic ints in
# bit-field offsets are part of the ISA encoding spec, not arbitrary.
"tools/asm/asm.py" = ["PLR0912", "PLR0915", "PLR2004"]
"tools/tinygrad_opengpu/codegen.py" = ["PLR2004"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------
[tool.mypy]
python_version = "3.10"
strict = true
warn_unreachable = true
disallow_untyped_defs = true
files = ["tools"]

[[tool.mypy.overrides]]
module = ["cocotb.*", "cocotb_bus.*", "cocotbext.*"]
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["test", "tools"]
addopts = "-ra --strict-markers --strict-config"
filterwarnings = [
    "error",
    "ignore::DeprecationWarning:cocotb.*",
]
markers = [
    "slow: long-running simulation tests",
    "fpga: requires an FPGA board attached",
    "synth: requires synthesis toolchain",
]

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------
[tool.coverage.run]
branch = true
source = ["tools", "test"]
omit = ["*/sim_build/*", "*/build/*"]

[tool.coverage.report]
show_missing = true
skip_covered = true
fail_under = 70
